matrix v2.0 (double precision)

+++ hill/matrix.h
@@ -14,16 +14,16 @@ extern "C" {
 #endif
 
     typedef struct{
-        int **data;
+        double **data;
         int size;
     }matrix;
 matrix *initMatrix(int size);
-void setValue(matrix *m,const int fila,const int col, const int value);
-int getValue(matrix *m,const int fila,const int col);
+void setValue(matrix *m,const int fila,const int col, const double value);
+double getValue(matrix *m,const int fila,const int col);
 void readMatrix(FILE *in, matrix *m);
 void printMatrix(matrix *m);
 void copyMatrix(matrix* src, matrix *dest);
-int detMatrix(matrix *m);
+double detMatrix(matrix *m);
 void eraseMatrix(matrix *m);
 void exchangeRows(matrix *m, int f1, int f2);
 #ifdef	__cplusplus
